+++ Program.cs
@@ -98,7 +98,8 @@ namespace MelonLoader
             {
                 string filePath = Directory.GetCurrentDirectory() + $@"\MLInstaller_{DateTime.Now:yy-M-dd_HH-mm-ss.fff}.log";
                 File.WriteAllText(filePath, msg);
-                Process.Start("explorer.exe", $"/select, {filePath}");
+                if (Config.HighlightLogLocation)
+                    Process.Start("explorer.exe", $"/select, {filePath}");
                 FinishingMessageBox($"INTERNAL FAILURE! Please upload the log file \"{filePath}\" when requesting support.", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
             catch (UnauthorizedAccessException)